+++ src/maidsafe/vault/pmid_manager/service.h
@@ -118,7 +118,7 @@ class PmidManagerService {
   template <typename Data>
   void HandlePutFailure(const typename Data::Name& data,
                         const PmidName& pmid_node,
-                        int32_t size,
+                        uint64_t size,
                         int64_t available_space,
                         const maidsafe_error& error_code,
                         nfs::MessageId message_id);
@@ -257,7 +257,7 @@ void PmidManagerService::HandlePut(const Data& data, const PmidName& pmid_node,
 
 template <typename Data>
 void PmidManagerService::HandlePutFailure(
-    const typename Data::Name& name, const PmidName& pmid_node, int32_t size,
+    const typename Data::Name& name, const PmidName& pmid_node, uint64_t size,
     int64_t available_space, const maidsafe_error& error_code, nfs::MessageId message_id) {
   LOG(kVerbose) << "PmidManagerService::HandlePutFailure to pmid_node -- "
                 << HexSubstr(pmid_node.value.string()) << " of size " << size
